+++ docs/changelog.md
@@ -4,6 +4,7 @@
 - *Effects* - Fixed issue where effects would get "stuck" on scenes after being played through Automated Animations
 - *Effects* - Fixed issue where the effect manager would break if opened with an effect that had pre-defined ranges in its `.file()` method 
 - *Crosshairs* - Fixed snap position resolution being undefined by default, causing it to not snap if not defined 
+- *Crosshairs* - Elaborated on some crosshairs function documentation  
 
 ## Version 3.5.0
 - *Sequencer* - Updated TJS implementation to latest versions