+++ docs/changelog.md
@@ -1,5 +1,9 @@
 ## Changelog
 
+## Version 3.5.2
+- *Effects* - Fixed `complete` effect versions not playing correctly without `.persist()`
+- *Effects* - Fixed isometric support with better plugin implementation
+
 ## Version 3.5.1
 - *Effects* - Fixed issue where effects would get "stuck" on scenes after being played through Automated Animations
 - *Effects* - Fixed issue where the effect manager would break if opened with an effect that had pre-defined ranges in its `.file()` method